requirements.txt
google-generativeai==0.8.4  # Ensure you use a valid version
Flask==2.2.5
google-cloud-storage==2.19.0
gunicorn==21.2.0
